# data/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ask HTTP API",
    "version": "0.1.0",
    "description": "Chat and ingest over files you indexed."
  },
  "paths": {
    "/chat": {
      "post": {
        "operationId": "chat",
        "summary": "Ask a question (JSON)",
        "description": "Blocking chat. Requires x-api-key. Optional source_types filter: markdown_docs, help_center, tickets, changelog, openapi.",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "description": "Must match API_KEY."
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": {
                    "type": "string"
                  },
                  "source_types": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "use_cache": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "question"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Answer, citations, and timings."
          },
          "401": {
            "description": "Invalid API key."
          }
        }
      }
    },
    "/chat/stream": {
      "post": {
        "operationId": "chatStream",
        "summary": "Ask a question (SSE)",
        "description": "Same body as POST /chat. Streams event: token then a final event: meta with citations.",
        "responses": {
          "200": {
            "description": "text/event-stream"
          },
          "401": {
            "description": "Invalid API key."
          }
        }
      }
    },
    "/ingest/run": {
      "post": {
        "operationId": "ingestRun",
        "summary": "Re-index files on disk",
        "description": "Full ingest pass. Unchanged files are skipped. Requires x-api-key.",
        "responses": {
          "200": {
            "description": "Counts of new, updated, unchanged, deleted."
          },
          "401": {
            "description": "Invalid API key."
          }
        }
      }
    },
    "/ingest/upload": {
      "post": {
        "operationId": "ingestUpload",
        "summary": "Upload and index documents",
        "description": "Multipart field `files`. Accepts UTF-8 .md, .txt, .html, .htm. Indexed immediately. Requires x-api-key.",
        "responses": {
          "200": {
            "description": "Per-file success or error."
          },
          "401": {
            "description": "Invalid API key."
          }
        }
      }
    },
    "/ingest/uploads": {
      "get": {
        "operationId": "listUploads",
        "summary": "List uploaded documents",
        "description": "Files saved under data/uploads/. Requires x-api-key.",
        "responses": {
          "200": {
            "description": "Filename, size, and updated_at."
          },
          "401": {
            "description": "Invalid API key."
          }
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "health",
        "summary": "Dependency health",
        "description": "Reports Qdrant, Redis, and configured LLM providers. No auth.",
        "responses": {
          "200": {
            "description": "Health payload."
          }
        }
      }
    }
  }
}